refactor(tui): split lib.rs event loop into event_loop module
lib.rs had grown to 943 lines (ADR 0028's watch threshold), mixing the
crate-root doc comment/mod declarations with run_app's event loop and
its dispatch machinery. Split by responsibility rather than a
mechanical top/bottom cut:

- event_loop/mod.rs: run_app itself, dispatch_non_source_key, the
  per-frame recompute gates, and the scroll-key classifier
- event_loop/scroll_sync.rs: ADR 0027/ADR 0030's diff-pane <-> tree
  cursor scroll synchronization, the hunk-jump target, and the
  post-draw scroll fold-back — grouped together since they all
  cooperate on the same right_pane_scroll state
- event_loop/goto.rs: ADR 0022's gd/gr candidate resolution

lib.rs itself is now a thin crate root (63 lines): mod declarations,
the crate-level doc comment, and ReviewPorts.

Also relocated three lib_tests/ files whose tests target functions
that already lived in sibling modules (input_translate.rs,
review_flow.rs) rather than lib.rs itself, following the same ADR
0028 pattern those modules' own doc comments already describe:
input_translate_tests/ and review_flow_tests/. Each test tree keeps
its own fixture helpers rather than sharing them across module
boundaries, matching the existing app/tests and detail_tests
precedent.

Rebased onto origin/main's #167 (which threaded a new DiffViewMode
argument through run_app/apply_diff_pane_selection_effects/
sync_target_for_scroll/dispatch_non_source_key/
auto_scroll_for_diff_focus): the same threading is ported into the
moved event_loop/ files and their tests, matching #167's change
exactly (verified against its diff site by site). 839 tests pass
(830 pre-split + 9 #167 added), matching origin/main's own count.
