--- rinkaku-tui/src/lib_tests/mod.rs
@@ -1,198 +1,0 @@
-//! Tests for `crate::lib` (the run/dispatch/translate glue), split from the
-//! source file to keep it under the ADR 0028 file-size threshold. Grouped
-//! by which crate-private function each set of tests pins:
-//!
-//! - `translate_key` — keyboard → `InputKey` translation, including the
-//!   help-overlay and jump-popup "swallow" contracts
-//! - `translate_mouse` — mouse-wheel/click translation
-//! - `recompute_and_reload` — the per-frame recompute gates
-//!   (`should_recompute_diff_pane_content`,
-//!   `should_recompute_blast_radius_selection`) and the source-cache
-//!   reload gate (`should_reload_source_content`)
-//! - `hunk_jump` — `should_apply_hunk_jump` and `jump_scroll_target`
-//! - `scroll_clamp` — post-draw fold-back
-//!   (`clamp_right_pane_scroll_after_draw`,
-//!   `clamp_help_scroll_after_draw`) and `is_scroll_input_key`
-//! - `goto_dispatch` — `resolve_goto` and the `dispatch_non_source_key`
-//!   regression coverage (gd/gr sequences, jumplist restore)
-//! - `diff_scroll_sync` — `sync_target_for_scroll` and
-//!   `apply_diff_pane_selection_effects` (ADR 0030): the diff → tree
-//!   auto-sync in the reverse direction of ADR 0027's tree → diff scroll
-//! - `perform_export` — `clipboard_export_status`'s OSC 52 size-guard
-//!   warning and error-message passthrough (ADR 0048)
-//! - `open_pr_in_browser` — the no-`PrContext`/spawn-failure status-line
-//!   messages and the URL built from a `PrContext` (ADR 0050)
-
-use crate::app::JumpCandidate;
-use crate::review::ports::BrowserOpener;
-use crate::source;
-use rinkaku_core::graph::SymbolGraph;
-use rinkaku_core::render::Report;
-
-mod diff_scroll_sync;
-mod goto_dispatch;
-mod hunk_jump;
-mod note_snapshot;
-mod open_pr_in_browser;
-mod perform_export;
-mod recompute_and_reload;
-mod scroll_clamp;
-mod translate_key;
-mod translate_mouse;
-
-/// A [`BrowserOpener`] fake shared by [`perform_export`]/[`open_pr_in_browser`]'s
-/// tests — `ReviewPorts::browser` is always present (ADR 0050), so every
-/// `ReviewPorts` fixture needs one even when the test itself is not
-/// exercising `w`. `opened_url` records the last URL passed to
-/// [`BrowserOpener::open_url`] so a test can assert the exact URL built from
-/// a [`crate::review::PrContext`], not just the resulting status message.
-pub(super) struct FakeBrowserOpener {
-    pub(super) result: Result<(), String>,
-    pub(super) opened_url: std::cell::RefCell<Option<String>>,
-}
-
-impl FakeBrowserOpener {
-    pub(super) fn new(result: Result<(), String>) -> Self {
-        Self {
-            result,
-            opened_url: std::cell::RefCell::new(None),
-        }
-    }
-}
-
-impl BrowserOpener for FakeBrowserOpener {
-    fn open_url(&self, url: &str) -> Result<(), String> {
-        *self.opened_url.borrow_mut() = Some(url.to_string());
-        self.result.clone()
-    }
-}
-
-pub(super) fn empty_report() -> Report {
-    Report {
-        origin: rinkaku_core::render::ReportOrigin::Diff,
-        files: vec![],
-        skipped: vec![],
-        graph: SymbolGraph {
-            nodes: vec![],
-            edges: vec![],
-            roots: vec![],
-        },
-        tests: vec![],
-        fan_ins: vec![],
-        file_size_warnings: vec![],
-        file_size_bands: vec![],
-        removed: vec![],
-    }
-}
-
-pub(super) fn report_with_one_symbol() -> Report {
-    use rinkaku_core::diff::LineRange;
-    use rinkaku_core::extract::{ExtractedSymbol, SymbolKind};
-    use rinkaku_core::render::FileReport;
-
-    Report {
-        files: vec![FileReport {
-            path: "lib.rs".to_string(),
-            symbols: vec![ExtractedSymbol {
-                id: "lib.rs::foo".to_string(),
-                name: "foo".to_string(),
-                kind: SymbolKind::Function,
-                signature: "fn foo()".to_string(),
-                range: LineRange { start: 1, end: 1 },
-                container: None,
-                referenced_names: vec![],
-                dependencies: vec![],
-                omitted_dependency_matches: 0,
-                is_test: false,
-                classification: None,
-                previous_signature: None,
-            }],
-        }],
-        ..empty_report()
-    }
-}
-
-pub(super) fn report_with_symbols_and_edges(
-    symbols_by_file: Vec<(&str, Vec<&str>)>,
-    edges: Vec<(&str, &str)>,
-) -> Report {
-    use rinkaku_core::diff::LineRange;
-    use rinkaku_core::extract::{ExtractedSymbol, SymbolKind};
-    use rinkaku_core::graph::{Edge, SymbolGraph};
-    use rinkaku_core::render::FileReport;
-
-    let files: Vec<FileReport> = symbols_by_file
-        .iter()
-        .map(|(path, names)| FileReport {
-            path: path.to_string(),
-            symbols: names
-                .iter()
-                .map(|name| ExtractedSymbol {
-                    id: format!("{path}::{name}"),
-                    name: name.to_string(),
-                    kind: SymbolKind::Function,
-                    signature: format!("fn {name}()"),
-                    range: LineRange { start: 1, end: 1 },
-                    container: None,
-                    referenced_names: vec![],
-                    dependencies: vec![],
-                    omitted_dependency_matches: 0,
-                    is_test: false,
-                    classification: None,
-                    previous_signature: None,
-                })
-                .collect(),
-        })
-        .collect();
-
-    let nodes: Vec<rinkaku_core::graph::Node> = symbols_by_file
-        .iter()
-        .flat_map(|(path, names)| {
-            names.iter().map(move |name| rinkaku_core::graph::Node {
-                id: format!("{path}::{name}"),
-                path: path.to_string(),
-                name: name.to_string(),
-                is_test: false,
-            })
-        })
-        .collect();
-
-    let graph_edges: Vec<Edge> = edges
-        .into_iter()
-        .map(|(from, to)| Edge {
-            from: from.to_string(),
-            to: to.to_string(),
-            is_cycle: false,
-        })
-        .collect();
-
-    Report {
-        files,
-        graph: SymbolGraph {
-            nodes,
-            edges: graph_edges,
-            roots: vec![],
-        },
-        ..empty_report()
-    }
-}
-
-pub(super) fn dummy_view(path: &str) -> source::HighlightedSourceView {
-    source::HighlightedSourceView {
-        view: source::SourceView {
-            path: path.to_string(),
-            lines: vec![],
-            highlight_start: 1,
-            highlight_end: 1,
-        },
-        token_highlights: vec![],
-    }
-}
-
-pub(super) fn candidate(id: &str, name: &str, path: &str) -> JumpCandidate {
-    JumpCandidate {
-        id: id.to_string(),
-        name: name.to_string(),
-        path: path.to_string(),
-    }
-}
